Create LinkedList.java
OUTPUT:
Enter the item to add to the list (or type 'done' to stop adding): apple
Add at (1) Beginning or (2) End? 2
Enter the item to add to the list (or type 'done' to stop adding): banana
Add at (1) Beginning or (2) End? 1
Enter the item to add to the list (or type 'done' to stop adding): cherry
Add at (1) Beginning or (2) End? 2
Enter the item to add to the list (or type 'done' to stop adding): done
Current Shopping List: Interns_CoreJavaAssignments_V001.LinkedList@7a07c5b4
Do you want to remove an item? (yes/no):